+++ src/index.ts
@@ -1,5 +1,5 @@
 export * from "./chains/ethereum";
 export * from "./mpcContract";
 export * from "./chains/near";
-export * from "./types";
+export * from "./types/types";
 export * from "./network";
